+++ x-pack/plugins/actions/common/index.ts
@@ -8,3 +8,5 @@
 export * from './types';
 
 export const BASE_ACTION_API_PATH = '/api/actions';
+
+export * from './rewrite_request_case';
