+++ x-pack/plugins/actions/server/routes/create.ts
@@ -9,9 +9,8 @@ import { schema } from '@kbn/config-schema';
 import { IRouter } from 'kibana/server';
 import { ActionResult, ActionsRequestHandlerContext } from '../types';
 import { ILicenseState } from '../lib';
-import { BASE_ACTION_API_PATH } from '../../common';
+import { BASE_ACTION_API_PATH, RewriteRequestCase, RewriteResponseCase } from '../../common';
 import { verifyAccessAndContext } from './verify_access_and_context';
-import { RewriteRequestCase, RewriteResponseCase } from './rewrite_request_case';
 import { CreateOptions } from '../actions_client';
 
 export const bodySchema = schema.object({
