+++ x-pack/plugins/actions/server/routes/execute.ts
@@ -10,10 +10,9 @@ import { IRouter } from 'kibana/server';
 import { ILicenseState } from '../lib';
 
 import { ActionTypeExecutorResult, ActionsRequestHandlerContext } from '../types';
-import { BASE_ACTION_API_PATH } from '../../common';
+import { BASE_ACTION_API_PATH, RewriteResponseCase } from '../../common';
 import { asHttpRequestExecutionSource } from '../lib/action_execution_source';
 import { verifyAccessAndContext } from './verify_access_and_context';
-import { RewriteResponseCase } from './rewrite_request_case';
 
 const paramSchema = schema.object({
   id: schema.string(),
