+++ x-pack/plugins/actions/server/routes/get_all.ts
@@ -7,10 +7,9 @@
 
 import { IRouter } from 'kibana/server';
 import { ILicenseState } from '../lib';
-import { BASE_ACTION_API_PATH } from '../../common';
+import { BASE_ACTION_API_PATH, RewriteResponseCase } from '../../common';
 import { ActionsRequestHandlerContext, FindActionResult } from '../types';
 import { verifyAccessAndContext } from './verify_access_and_context';
-import { RewriteResponseCase } from './rewrite_request_case';
 
 const rewriteBodyRes: RewriteResponseCase<FindActionResult[]> = (results) => {
   return results.map(({ actionTypeId, isPreconfigured, referencedByCount, ...res }) => ({
