+++ x-pack/plugins/triggers_actions_ui/public/application/lib/action_connector_api/delete.ts
@@ -1,0 +1,28 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+import { HttpSetup } from 'kibana/public';
+import { BASE_ACTION_API_PATH } from '../../constants';
+
+export async function deleteActions({
+  ids,
+  http,
+}: {
+  ids: string[];
+  http: HttpSetup;
+}): Promise<{ successes: string[]; errors: string[] }> {
+  const successes: string[] = [];
+  const errors: string[] = [];
+  await Promise.all(ids.map((id) => http.delete(`${BASE_ACTION_API_PATH}/connector/${id}`))).then(
+    function (fulfilled) {
+      successes.push(...fulfilled);
+    },
+    function (rejected) {
+      errors.push(...rejected);
+    }
+  );
+  return { successes, errors };
+}
