+++ x-pack/plugins/triggers_actions_ui/public/application/lib/action_connector_api/index.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { loadActionTypes } from './connector_types';
+export { loadAllActions } from './connectors';
+export { createActionConnector } from './create';
+export { deleteActions } from './delete';
+export { executeAction } from './execute';
+export { updateActionConnector } from './update';
