+++ x-pack/plugins/triggers_actions_ui/public/application/lib/action_connector_api/update.test.ts
@@ -1,0 +1,49 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { ActionConnectorWithoutId } from '../../../types';
+import { httpServiceMock } from '../../../../../../../src/core/public/mocks';
+import { updateActionConnector } from './index';
+
+const http = httpServiceMock.createStartContract();
+
+beforeEach(() => jest.resetAllMocks());
+
+describe('updateActionConnector', () => {
+  test('should call the update API', async () => {
+    const id = '123';
+    const apiResponse = {
+      connector_type_id: 'test',
+      is_preconfigured: false,
+      name: 'My test',
+      config: {},
+      secrets: {},
+      id,
+    };
+    http.put.mockResolvedValueOnce(apiResponse);
+
+    const connector: ActionConnectorWithoutId<{}, {}> = {
+      actionTypeId: 'test',
+      isPreconfigured: false,
+      name: 'My test',
+      config: {},
+      secrets: {},
+    };
+    const resolvedValue = { ...connector, id };
+
+    const result = await updateActionConnector({ http, connector, id });
+    expect(result).toEqual(resolvedValue);
+    expect(http.put.mock.calls[0]).toMatchInlineSnapshot(`
+      Array [
+        "/api/actions/connector/123",
+        Object {
+          "body": "{\\"name\\":\\"My test\\",\\"config\\":{},\\"secrets\\":{}}",
+        },
+      ]
+    `);
+  });
+});
