+++ x-pack/test/functional_with_es_ssl/apps/triggers_actions_ui/alerts_list.ts
@@ -42,7 +42,7 @@ export default ({ getPageObjects, getService }: FtrProviderContext) => {
 
   async function createAction(overwrites: Record<string, any> = {}) {
     const { body: createdAction } = await supertest
-      .post(`/api/actions/action`)
+      .post(`/api/actions/connector`)
       .set('kbn-xsrf', 'foo')
       .send(getTestActionData(overwrites))
       .expect(200);
